refactor(pr3): add public facade and migrate CLI consumer
