+++ packages/newsletter_core/src/newsletter_core/public/lifecycle.py
@@ -1,0 +1,17 @@
+"""Public lifecycle primitives for runtime adapters."""
+
+from __future__ import annotations
+
+from newsletter.utils.shutdown_manager import (
+    ShutdownPhase,
+    get_shutdown_manager,
+    is_shutdown_requested,
+    register_shutdown_task,
+)
+
+__all__ = [
+    "ShutdownPhase",
+    "get_shutdown_manager",
+    "is_shutdown_requested",
+    "register_shutdown_task",
+]
